Print formatted names for parents

Displaying names of parent taxa.

Suffixes in the function names are explained as follows:

pfn: Parent's full name without author name.

pfna: Parent's full name with author name.

pfnas: Parent's full name with author name and taxon view (secundum).

pan: Parent's abbreviated name without author name.

pana: Parent's abbreviated name with author name.

tn_pfn(x, level, ...)

tn_pfna(x, level, ...)

tn_pfnas(x, level, ...)

tn_pan(x, level, ...)

tn_pana(x, level, ...)

Arguments

x

A vector containing selected identifiers for taxon concepts (TaxonConceptID) included in the internal taxonomic object (see set_tax()).

level

A character value indicating the taxon rank set for parent taxa. I is passed to taxlist::parents().

...

Further arguments passed to taxlist::print_name().

Value

All these functions return a character value.

Examples

## Compare all these outputs
tn_pfn(13, "family")
#> [1] "*Leguminosae*"
tn_pfna(13, "family")
#> [1] "*Leguminosae* Juss."
tn_pfnas(13, "family")
#> [1] "*Leguminosae* Juss. sec. The Plant List (2013)"
tn_pan(13, "family")
#> [1] "*Leguminosae*"
tn_pana(13, "family")
#> [1] "*Leguminosae* Juss."